Add: verify same user or admin for delete function

+++ src/controllers/reservation.controller.js
@@ -120,6 +120,12 @@ const deleteReservationById = async (req, res) => {
             return res.json({ message: "Reserva no encontrada" })
         }
 
+         //es del mismo usuario q eliminara su reserva? o admin?
+        if(reservation.user.toString() !== req.user.id && !req.user.isAdmin){
+            res.status(403)
+            return res.json({message: "No autorizado"})
+        }
+        
        
         const filters = { _id: req.params.id }
         await Reservation.deleteOne(filters)
